Restructure proxy to be a private portion of the network driver
